content length in multiparse context object adjusted to match the input type size (LL instead of L)

+++ src/multipart.c
@@ -51,7 +51,7 @@ THE SOFTWARE.
 
 typedef struct {
   char *lbuf2; /* used for file reading as line tail */
-  long content_length; /* currently unused */
+  long long content_length; /* currently unused */
   int error;
   cstr_t buf;
   char *boundary;
